Allow the user to add a new book to the array
Make it so a book with the data provided by the user through the dialog
is added to the library array and that it's also displayed on the page.
Add an attribute to the constructor of the Book object that keeps
track of whether the book is already being displayed. this
attribute is initially set to 0 (not displayed) and it gets changed to
a 1 after it gets displayed; this is done so that it doesn't get
displayed multiple times. Add an if to check this value in the
method that displays the books.